feat(velocity): adding base64 encoding/decoding
ref: #33216
